treewide: add support for 80+80 MHz channel
OpenWrt's iwinfo exposes many values like VHT80_80 [1] and HE80_80 [2],
but while testing miniPCIe card Compex WLE1216V5 [3] inserted to Turris
Omnia router, I noticed that in LuCI, it is possible to configure
channel width '20 MHz', '40 MHz', '80 MHz' and '160 MHz', but according
to the datasheet [4], this miniPCie card supports 80+80 MHz and not 160 MHz, which is different.
It avoid DFS channels and operate on two channels simultanously.

Before this commit:
- LuCI configures this card by looking at uci as "htmode": "VHT160"

After this commit:
- LuCI

-- commit message needs to be improved --

[1] https://git.openwrt.org/?p=project/iwinfo.git;a=blob;f=include/iwinfo.h;h=8469ee70f5e36342780817236d3994a440e0788e;hb=HEAD#l91
[2] https://git.openwrt.org/?p=project/iwinfo.git;a=blob;f=include/iwinfo.h;h=8469ee70f5e36342780817236d3994a440e0788e;hb=HEAD#l97
[3] Network controller [0280]: Qualcomm Atheros QCA9984 802.11ac Wave 2 Wireless Network Adapter [168c:0046]
[4] https://compex.com.sg/wp-content/uploads/2021/01/wle1216v5-20-v1.3-sl-227.pdf

Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>